decouple host and host builder
