+++ src/SampSharp.OpenMp.Entities/Hosting/EcsBuilder.cs
