+++ src/SampSharp.OpenMp.Entities/Hosting/IEcsStartup.cs
