+++ src/SampSharp.OpenMp.Entities/Hosting/SampSharpEnvironment.cs
