+++ src/SampSharp.OpenMp.Entities/Hosting/UnhandledExceptionHandler.cs
