+++ src/TestMode.OpenMp.Entities/Startup.cs
@@ -10,11 +10,9 @@ public class Startup : IEcsStartup
 {
     public void Initialize(IStartupContext context)
     {
-        context.UseEntities(cfg =>
-        {
-            cfg.UsePlayerCommands();
-            cfg.ConfigureLogging(logging => logging.SetMinimumLevel(LogLevel.Information));
-        });
+        context.UseEntities()
+            .UsePlayerCommands()
+            .ConfigureLogging(logging => logging.SetMinimumLevel(LogLevel.Information));
     }
 
     public void ConfigureServices(IServiceCollection services)
